Slight Cleanup. Libre Cell Implementation Added.

+++ include/boost/document/detail/cell_interface.hpp
@@ -13,7 +13,7 @@ namespace boost {
 	class cell_interface {
 	 public:
 	 	virtual void set_cell_value(const std::string& str) = 0;
-    	virtual void set_cell_value(int i) = 0;
+    	virtual void set_cell_value(float i) = 0;
    		virtual ~cell_interface() {}
 	};	
 } // namespace boost
